fix :: DTO 변환

+++ src/main/kotlin/xquare/app/xquareinfra/domain/container/application/service/UpdateEnvironmentVariableService.kt
@@ -60,7 +60,7 @@ class UpdateEnvironmentVariableService(
         pipelinesHistory.pipelines?.get(0)?.let {
             gocdClient.runSelectedJob(
                 pipelineName,
-                it.counter as Int,
+                it.counter,
                 "deploy",
                 accept,
                 RunSelectedJobRequest(jobs = listOf("deploy"))
